refactor api manager into multiple files
