+++ src/components/LoggedInNavbar.js
@@ -2,7 +2,7 @@ import React, { useContext, useState } from 'react'
 import { useHistory, useLocation } from 'react-router-dom'
 import UserContext from '../context/UserContext'
 import { LOGOUT_USER } from '../actionTypes'
-import LoginManager from '../helpers/LoginManager'
+import LoginManager from '../apiClients/LoginManager'
 import Avatar from '../subcomponents/Avatar'
 import Dropdown from '../subcomponents/Dropdown'
 import InvalidPaymentBanner from '../subcomponents/InvalidPaymentBanner'
