+++ src/components/without_auth_flow/Signup.js
@@ -5,7 +5,7 @@ import Input from '../../subcomponents/Input'
 import Submit from '../../subcomponents/Submit'
 import Form from '../../subcomponents/Form'
 import WithFooter from '../../wrappers/WithFooter'
-import LoginManager from '../../helpers/LoginManager'
+import LoginManager from '../../apiClients/LoginManager'
 import { POPULATE_USER, LOADING, LOGOUT_USER } from '../../actionTypes'
 import UserContext from '../../context/UserContext'
 import { Link } from 'react-router-dom'
